+++ example/lib/pages/card.dart
@@ -71,7 +71,7 @@ class _CardPageState extends State<CardPage> {
               crossAxisAlignment: CrossAxisAlignment.stretch,
               children: [
                 const Text('Name'),
-                const ShadInput(placeholder: Text('Name of your project')),
+                const ShadInput(hintText: 'Name of your project'),
                 const SizedBox(height: 6),
                 const Text('Framework'),
                 ShadSelect<String>(